requirements.txt
# 核心依赖
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0

# 爬虫相关
playwright>=1.40.0
beautifulsoup4>=4.12.0
fake-useragent>=1.4.0

# AI 与 LLM 相关
openai>=1.3.0
langchain>=0.0.339

# 数据库与工具
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.2

